requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-phonenumber-field==5.0.0
flake8==3.8.4
gunicorn==20.0.4
mccabe==0.6.1
phonenumbers==8.12.11
Pillow==8.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.4.1
